Use enum for cardSize
