+++ src/components/ui/WorkCard.astro
@@ -1,6 +1,7 @@
 ---
 import { Image } from "astro:assets";
-import type { WorkType } from "@/content.config";
+
+import { CardSize, type WorkType } from "@/schema/work";
 
 interface Props {
   work: {
@@ -12,11 +13,21 @@ interface Props {
 const {
   work: {
     id,
-    data: { title, description, type, tech, link, startDate, cardSize, cover, coverAlt = "" },
+    data: {
+      title,
+      description,
+      type,
+      tech,
+      link,
+      startDate,
+      cardSize,
+      cover,
+      coverAlt = "",
+    },
   },
 } = Astro.props;
 
-const isSmallCard = cardSize === "small";
+const isSmallCard = cardSize === CardSize.SMALL;
 const path = `/work/${id}/`;
 ---
 
@@ -27,7 +38,11 @@ const path = `/work/${id}/`;
         <div class="bg-gray-200 w-full">
           {cover && (
             <a href={path} class="block">
-              <Image src={cover} alt={coverAlt} class="aspect-400/240 object-cover object-top" />
+              <Image
+                src={cover}
+                alt={coverAlt}
+                class="aspect-400/240 object-cover object-top"
+              />
             </a>
           )}
         </div>
